Add support for different ipmitool privilege levels.
Also need to add support for separate ipmi encryption keys from user
passwords, but it turns out that the current db schema does actually
support that already and at least one script more or less does it
already (power_ipmi.pm), though another doesn't (power_ilo.pm), so we
can fix that separately there.

This is to allow support of additional read-only users for gathering
power data that don't have to also be given the admin password in form
form of the hex encoded Kg key.

Also add a field so that we don't need to assume that the privilege
level is necessarily ADMINISTRATOR.