Add AddNodes and DeleteNodes, which are convenience functions for the HPC folks.
AddNodes($slice_urn, $credentials, $nodes):

The "nodes" argument is a hash that looks like:

  {"node45" : {"diskimage" : "urn...",
               "startup"   : "/bin/echo",
               "tarballs"  : ["tarball1", "tarball2", ...],
               "lans"      : ["lan1", "lan2", ...]
               "node"      : "pc189"},
   "nodeXX" : {...}}

DeleteNodes($slice_urn, $credentials, $nodes):

The "nodes" argument is a list like:

  ["node45", ...]

Any node can be deleted, but it is not yet clear what happens if all the
nodes of a lan are removed. I probably need to do some work there, but
David can start with this.