New sitevar to set a default per-project dataset quota.
In createdataset, if the "usequotas" sitevar is set for the dataset type in
question but a quota does not exist for the dataset's project, we create
a quota object using the value from the new "default_quota" sitevar for that
dataset type. If that sitevar does not exist or has a value of zero, we do
NOT create a quota object and hence createdataset will fail.