Rework how OpenFlow setup is handled on the front-end
Previously OF would be setup after vlan creation, but before checking to
see if all switches in scope actually support it.  This would lead to
failures where the vlan was created and logged in the DB, but snmpit
failed to setup OF, and so exit with a failure.  This in turn causes
swap-in to fail, and leaves stale state on the switches and in the DB.

Now when OF is requested for a vlan in an experiment, we first check
to see if it is supported, then proceed.  We error out early before
any OF-enabled vlan is created.