Add support for OpenFlow to Force10 snmpit module.
Adding this support required a change to the createVlan() API call,
which is why this commit touches so many modules. Optional/supplementary
arguments to createVlan() are now passed as a fourth hash reference
argument.

This supplementary hash is used to inform device modules that a vlan to
be created is to be associated with an OpenFlow instance.  Only the
snmpit_force10 makes use of this information at this time. FTOS requires
that openflow-enabled vlans be declared as such at creation time. However,
snmpit's default control flow creates vlans first, then associates them
with OF instances (if requested).  This is why the supplementary argument
(OF hint) to createVlan() is needed here.

OpenFlow configuration actions must all be performed via the
existing Expect-based CLI interface to FTOS. Yuck.