Add password block decryption and expansion in the instructions panel. Given a password element in the rspec:
	<emulab:password name='foo'></password>

which the portal has converted to an encrypted secret, when that experiment
is later shown (the status page), ask the server to decrypt the block, and
then replace the string "{password-foo}" in the instructions with the
actual password.

Need to generalize this a bit more, for arbitrary encryption blocks, when
we have those.