Experimental PACK_TIGHT compile time option
When set, the goal is to pack vnodes as tightly as possible - that is,
to penalize wasted slots in assignments.

One downside is that this will tend to make really lopsided mappings:
eg. if you have 50 slots per pnode, and 51 vodes, you will end up
mapping 50 to one and 1 to the other.

The other downside is that I have not yet thought about how to balance
this with link weights.