Make systemd swap hack run on Ubuntu 15.
So, the scripts are now in clientside/tmcc/linux since they're the same
for any systemd-based Linux, and the centos7 and ubuntu15 makefiles
install those files.

Also had to improve fixup-fstab-swaps to handle the case where the / is
something like /dev/disk/by-uuid/<UUID> which is a symlink to the real
dev.  So try readlink -f to handle that.