Fix bit rot when deleting addresses in addvpubaddr.
It assumed addresses were unused if node_id was NULL.  This used to be
true before the introduction of per-experiment dynamic address pools (when
public addresses were mapped one-to-one with VMs).  Unfortunately this
code was not updated when experiment address pools were added, and so
it would have permitted the deletion of currently assigned addresses.