Extract select DB info about node allocations over time.
For a date range, return records for all nodes allocated during that
period along with what project/experiment they were in, what users are
associated with the project and info about those users.

I added an option to anonymize the data, but I didn't implement it...
Little ditty written to extract node info for students in Kobus's
networking class.