Systemd units for testbed and tbprepare; consistent systemd install.
Currently ubuntu15 and centos7 are our only systemd-based OSes; they
both call into the clientside/tmcc/linux generic Makefile with the
right magic bits to get the systemd sysetc-install/remove targets.
For now, systemd is effectively the same on centos7 or ubuntu15, so
we use the same scripts and install strategy from tmcc/linux for
both (except for ntpd on centos7, which stays as it was).

I used the same testbed/tbprepare sysv init scripts in tmcc/linux,
but I installed them in BINDIR/initscripts in a lame attempt to
avoid confusion ;).  I modified the testbed script slightly so
it looks for a magic file in /etc/emulab/(uses-systemd) to know
that systemd will start the testbed service, and that it should
not get started from rc.local .  So those lines are now back in
the generic linux testbed script -- just unused if the init is
not systemd (in which case rc.local starts it).

These units do print out to both the systemd journal and to the
console.