Small addition to allow easier prereserves for nonlocal projects (created via the Geni API, which have hard to guess project names).
	boss> wap perl prereserve emulab.net:testbed 1

At the remote cluster, you prepend the domain of the SA (for the portal,
that is emulab.net) to the project name at the portal.  prereserve will
find the correct project, assuming at least one experiment has ever been
created on the remote cluster. If not, then this trivial little hack will
not work.