When doing a "reset", reverse the order of the scripts.
This is when we are cleaning up to make an image. Seems like the correct
thing to do is process the startup scripts in reverse order like we do for
a "shutdown". Mostly this doesn't matter, but it does matter for the
blockstore cleanup where we need to process the remote BSes before the locals.