Make "reset" operation work for blockstores.
Reset on local/remote blockstores ensures that there is no blockstore related
state left in the root filesystem (e.g., mounts in /etc/fstab, iSCSI config,
LVM/ZFS state). It does this in such a way that upon reboot, all the necessary
state is recreated.

What this means is that you should now be able to take an image of a node
that uses blockstores and have that image actually work on another node!
Previously, there could/would be leftover blockstore turds that would make
the new image fail to boot.

Of course, this won't work until the standard images are remade and will
then only work for those images or images derived from them.