Fix up nav bar for narrow (<= 768) screen, use more of the bootstrap responsiveness stuff to make the shortcut buttons move out of the navbar and into the menu. Also move the admin stuff into its own menu.
This was a lot harder then it needed to be, but I did not like the standard
bootstrap navbar collapse into hamburger menu approach, it looked terrible
with our header, and takes up a TON of vertical space for no reason I can
see. I prefer the consistent menu display, regardless of screen width, so I
had to make a few tweaks to the dropdown class; bootstrap changes a bunch
of settings in a media clause, which makes the menus look stupid, again,
for no reason that I can see.