New (test) version of image_import that can import the entire image history from the server, keeping it in sync with the server as new versions of the image are added. Also handles importing deltas if the metadata says there is a delta.
Note that downloading the image files is still lazy; we will not import all
15 versions of an image unless they actually are needed.

Lots of work still do. This is a bit of a nightmare cause of client/server
(backward) compatibility issues wrt provenance/noprovenance and
deltas/nodeltas. I might change my mind and say the hell with
compatibility!

Along these same lines, there is an issue of what to do when a site that is
running with provenance turned on, gets this new code. Up to now, the
client and server never tried to stay in sync, but now they have to (cause
of deltas), and so the client image descriptors have to be upgraded. That
will be a hassle too.