tbadb_serv: separate port forwarding reservation from activation.
Since we need to determine the port adb will listen on for a device
early on in the experiment swap-in process (right after mapping), we
call an RPC to just reserve the port first.  Later on during experiment
resource setup, we call an "activate" RPC to actually setup the adbd
port forwarding.

I also fixed up error_exit() to log its caller as the function that had
the error, not itself.  It uses the Net::Server class' log() method now
as well.