Don't do systemctl call if DESTDIR is set.
I want to be able to make a client tarball. Actually, I would prefer that
this call was not here at all or was at least the very last thing in a
client install. When I did a build on CentOS 7, the make hung here everytime
before timing out and leaving me with a partial install.

But I ain't gonna touch it...systemd cooties!