You can now interact with grub on the d430s!
Apparently, specifying "--unit==1" to grub did not work because grub
didn't think it had a unit 1 (aka, COM2), so it rejected the "serial"
command. Output still worked due to redirection (I think), but it was
expecting input from the keyboard.

It looks like the number of valid serial ports is compiled into grub and
who knows how it was built for Ubuntu. Anyway, by specifying "--port=0x2F8"
in addition, that will override the unit test and all is well. Strange.

Also enhanced the Linux slicefix to deal with fixing /etc/default/grub so
that those wacky console settings will remain in effect even if grub is
re-configured (which some packages do).