Progmode tweaks to address our recent hostile takeover of boss by capture.
If the app fails to launch after 10 tries, give up.

Refined the hack to detect the case where we wind up in progmode()
on a restart before the old child has completely disappeared.
Somewhere we still have a race that leaves behind capture zombies.
The party really starts when we run out of processes!