Modify to include two additional user-related fields.
Add a "last activity" timestamp indicating the last time the user was
known to interact with Emulab by logging into the portal, logging into
a node or performing some other activity that resulted in their home
directory being exported (this latter may be dumb, not sure yet).
This should provide additional information about whether a user is really
using Emulab or is just someone who was in a project once and maybe
graduated and moved on.

Added a canonical "location" string to identify the user's home city.
This is the "formatted name" returned by the Google Maps API call
(with commas replaced by semicolons to prevent issues with CSV export
of the data). This field is intended to be a canonical rendition of the
city/state/country info the user enters.

That script was also modified to include users that are "frozen" or
"archived" by default. Unfortunately, just because they are in those
states now, does not mean they were during the desired time range.
Emulab keeps no indication of when they might have changed state.

The new -U option can be used to elide records for users that we have
determined were not active during the time range (i.e., their "last
activity" timestamp above is before the indicated start time.