Add Ubuntu 16 support (both x86 and aarch64).
This is all pretty vanilla atop the Ubuntu 15 support.  Just new
metadata files (passwd,group,shadow,gshadow); pulling in Kirk's Centos
ntp.service; and an important tweak to the dhclient exit hook.

resolvconf seems to be more prominently featured in Ubuntu 16, and it
has a dhclient exit hook to alert the resolvconf stuff (so it can alert
others) when nameservers change.  However, that hook somehow breaks DNS
for our exit hook right as it calls sethostname.dhclient .  So to fix
this, I just added a check for /sbin/resolvconf in the exit hook, and if
that exists, we use the same elabinelab DNS-problem-avoidance strategy
of manually invoking the tmcc binary with getbossinfo, so that the perl
tmcc stuff doesn't hang in sethostname.client .  That works great.

aarch64 (ubuntu16-ms) support is a very simple overlay atop the regular
ubuntu16 dir; it only adds the special uboot boot.scr modification file
we need.