This change is to support switching urn in data structure from strings to objects (see GeniHRN, look for new()). Much easier, less typing.  But in order to do that, we have make sure that if we send one on the wire, it gets converted properly (converted to it its plain string).
But, Frontier and XML::RPC do not let you hook in so that if you have a
blessed reference, it will call its stringify method for encoding.  Odd,
cause encode_json (JSON) supports that.

As it turns out, Frontier is structured so that it is easy to hook into
it, and its only mildly sleazy. And Frontier has not changed in years,
so probably not going to change much in the next few years. XML::RPC was
too messy, so I switched Genixmlrpc CallMethod() to use Frontier
instead.  Does not seem to risky, we use Frontier on the receiving end
anyway.

Lets see how this goes, its been running in my devel tree for a while.