Allow the systemd fstab swap fixer to mkswap/swapon!
I cannot find why we called fixup-fstab-swaps with '-E' (which means,
don't mkswap/swapon any swaps).  The only thing I can think of is that
perhaps running swapon manually made the systemd dev-*.swap targets
unhappy.  However, it is necessary to mkswap if the swap device didn't
exist, because systemd will not mkswap for you, AFAIK -- it will only
swapon.  On Ubuntu 16, the dev-*.swap targets are happy whether they or
Emulab does the swapon.  If that's not true on Centos 7 or other
systemds, we may have to revisit this tweak.